+++ src/proxy/src/calls/websocket_calls.rs
@@ -1,5 +1,5 @@
 use candid::Principal;
-use canister_types::models::websocket_message::WSMessage;
+use catalyze_shared::websocket_message::WSMessage;
 use ic_cdk::{query, update};
 use ic_websocket_cdk::{
     CanisterWsCloseArguments, CanisterWsCloseResult, CanisterWsGetMessagesArguments,