+++ src/proxy/src/helpers/mod.rs
@@ -1,6 +1,3 @@
 pub mod group_permission;
 pub mod guards;
-pub mod str;
-pub mod time_helper;
 pub mod token_balance;
-pub mod validator;