+++ src/proxy/src/logic/attendee_logic.rs
@@ -1,5 +1,5 @@
 use candid::Principal;
-use canister_types::models::{
+use catalyze_shared::{
     api_error::ApiError,
     attendee::{Attendee, JoinedAttendeeResponse},
 };