+++ src/proxy/src/logic/friend_request_logic.rs
@@ -1,5 +1,5 @@
 use candid::Principal;
-use canister_types::models::{
+use catalyze_shared::{
     api_error::ApiError,
     friend_request::{FriendRequest, FriendRequestResponse},
     profile::ProfileResponse,