+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `catalyze-shared` dependency
+
+### Removed
+
+- crate `canister_types`
+- string, type, validation helpers
+
 ## [0.2.4]
 
 ### Added
@@ -355,7 +366,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Missing `notification_id` on migration models
 
-[Unreleased]: https://github.com/Catalyze-Software/proxy/compare/0.2.2...HEAD
+[Unreleased]: https://github.com/Catalyze-Software/proxy/compare/0.2.4...HEAD
+[0.2.4]: https://github.com/Catalyze-Software/proxy/compare/0.2.3...0.2.4
+[0.2.3]: https://github.com/Catalyze-Software/proxy/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/Catalyze-Software/proxy/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/Catalyze-Software/proxy/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/Catalyze-Software/proxy/compare/0.1.9...0.2.0